+++ docs/logs/redaction/MyRedactionProcessor.cs
@@ -18,11 +18,19 @@ using System.Collections.Generic;
 using OpenTelemetry;
 using OpenTelemetry.Logs;
 
+namespace Redaction;
+
 internal class MyRedactionProcessor : BaseProcessor<LogRecord>
 {
     public override void OnEnd(LogRecord logRecord)
     {
-        if (logRecord.State is IReadOnlyList<KeyValuePair<string, object>> listOfKvp)
+        if (logRecord.State == null)
+        {
+            // When State is null, OTel SDK guarantees StateValues is populated
+            // TODO: Debug.Assert?
+            logRecord.StateValues = new MyClassWithRedactionEnumerator(logRecord.StateValues);
+        }
+        else if (logRecord.State is IReadOnlyList<KeyValuePair<string, object>> listOfKvp)
         {
             logRecord.State = new MyClassWithRedactionEnumerator(listOfKvp);
         }