+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.4.0-beta.2
+
+Released 2022-Oct-17
+
 * OTLP histogram data points will now include `Min` and `Max` values when
   they are present.
   ([#2735](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2735))
@@ -11,6 +15,10 @@
   [Attribute Limits](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/sdk-environment-variables.md#attribute-limits)
   are configured via the environment variables defined in the specification.
   ([#3684](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3684))
+
+* Added support for loading environment variables from `IConfiguration` when
+  using the `AddOtlpExporter` extensions
+  ([#3760](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3760))
 
 ## 1.4.0-beta.1